pyproject.toml
[project]
name = "cei"
version = "0.1.0"
requires-python = ">=3.11,<3.14"
dependencies = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "rouge-score==0.1.2",
    "sqlitedict==2.1.0",
    "word2number==1.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
cloud = [
    "boto3>=1.34",                  # AWS
    "google-cloud-compute>=1.18",   # GCP
    "azure-identity>=1.16",         # Azure auth
    "azure-mgmt-compute>=31.0",     # Azure VMs
    "azure-mgmt-network>=25.0",     # Azure networking
    "azure-mgmt-resource>=23.0",    # Azure resource groups
]

[project.scripts]
cei = "src.infra:main"

[tool.uv.workspace]
members = ["src/lm-evaluation-harness", "src/inspect"]

[tool.uv]
find-links = ["vendor/"]

[dependency-groups]
dev = ["pytest>=8.0", "pyyaml>=6.0", "datasets>=2.19", "inspect-ai>=0.3"]
